+++ src/rt_door.c
@@ -89,7 +89,7 @@ boolean areabyplayer[NUMAREAS];
 
 // Local Variables
 
-static void (*touchactions[NUMTOUCHPLATEACTIONS])(long) = {
+static void (*touchactions[NUMTOUCHPLATEACTIONS])(intptr_t) = {
 	ActivatePushWall, ActivateMoveWall, LinkedOpenDoor, LinkedCloseDoor,
 	EnableObject,	  DisableObject,	ActivateLight,	DeactivateLight};
 
@@ -273,7 +273,7 @@ void DoAnimatedMaskedWalls(void)
 	}
 }
 
-int GetIndexForAction(void (*action)(long))
+int GetIndexForAction(void (*action)(intptr_t))
 {
 	int i;
 
@@ -400,11 +400,11 @@ void LoadTouchPlates(byte* buffer, int size)
 		temp->complete = dummy.complete;
 
 		if (dummy.whichobj & FL_TACT)
-			temp->whichobj = (long)(objlist[dummy.whichobj & ~FL_TACT]);
+			temp->whichobj = (intptr_t)(objlist[dummy.whichobj & ~FL_TACT]);
 
 		else if (dummy.whichobj & FL_TSTAT)
 			temp->whichobj =
-				(long)(GetStatForIndex(dummy.whichobj & ~FL_TSTAT));
+				(intptr_t)(GetStatForIndex(dummy.whichobj & ~FL_TSTAT));
 		else
 			temp->whichobj = dummy.whichobj;
 		if (dummy.actionindex != -1)
@@ -473,8 +473,8 @@ void RemoveTouchplateAction(touchplatetype* tplate, int index)
 	totalactions--;
 }
 
-void Link_To_Touchplate(word touchlocx, word touchlocy, void (*maction)(long),
-						void (*swapaction)(long), long wobj, int delaytime)
+void Link_To_Touchplate(word touchlocx, word touchlocy, void (*maction)(intptr_t),
+						void (*swapaction)(intptr_t), intptr_t wobj, int delaytime)
 {
 	touchplatetype* temp;
 	int index;
@@ -501,7 +501,7 @@ void Link_To_Touchplate(word touchlocx, word touchlocy, void (*maction)(long),
 	totalactions++;
 }
 
-void ClockLink(void (*saction)(long), void (*eaction)(long), long wobj,
+void ClockLink(void (*saction)(intptr_t), void (*eaction)(intptr_t), intptr_t wobj,
 			   int whichclock)
 {
 	touchplatetype* temp;
@@ -592,7 +592,7 @@ void TriggerStuff(void)
 	touchplatetype* temp;
 	int i, touchcomplete, j;
 	int playeron;
-	void (*tempact)(long);
+	void (*tempact)(intptr_t);
 	boolean wallmessage, doormessage, columnmessage;
 
 	for (i = 0; i < lasttouch; i++)
@@ -1664,7 +1664,7 @@ void OperateDoor(int keys, int door, boolean localplayer)
 =====================
 */
 
-void LinkedOpenDoor(long door)
+void LinkedOpenDoor(intptr_t door)
 {
 	UtilizeDoor(door, OpenDoor);
 }
@@ -1677,7 +1677,7 @@ void LinkedOpenDoor(long door)
 =====================
 */
 
-void LinkedCloseDoor(long door)
+void LinkedCloseDoor(intptr_t door)
 {
 	if (DoorReadyToClose(door) == true)
 		UtilizeDoor(door, CloseDoor);
@@ -3286,7 +3286,7 @@ void ActivateAllPushWalls(void)
 =====================
 */
 
-void ActivatePushWall(long pwall)
+void ActivatePushWall(intptr_t pwall)
 {
 	pwallobj_t* pw;
 
@@ -3317,7 +3317,7 @@ void ActivatePushWall(long pwall)
 =====================
 */
 
-void ActivateMoveWall(long pwall)
+void ActivateMoveWall(intptr_t pwall)
 {
 	pwallobj_t* pw;
